pitch-shift/requirements.txt
ffmpeg-python
librosa
pytube
yt-dlp
